Allow easy copy/paste of quiz questions
Anyone with access to quizzes/<QUIZ_NUMBER>/edit will be able to make use of this short script. When the script is enabled with TamperMonkey, it will create an API call on page load and populate a modal. Clicking the created button will allow the modal to be opened, and the content copied in an easy to read format.